# frontend/src/generated/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Starlib API",
    "description": "Backend API for Starlib music management",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/soundcloud/tracks/{track_id}/stream": {
      "get": {
        "tags": [
          "soundcloud"
        ],
        "summary": "Get Track Stream",
        "description": "Return a signed HLS playlist URL for the given SoundCloud track.\n\nThe result is cached in-memory and reused while still valid. Expired\nentries are transparently refetched.\n\nParameters\n----------\ntrack_id : int\n    SoundCloud track id.\nforce_refresh : bool, optional\n    When true, evict any cached entry and fetch a fresh URL from\n    SoundCloud. Clients should set this after receiving a 403 on a\n    cached URL so the server does not hand them the same stale entry.\n\nReturns\n-------\nStreamUrlResponse\n    ``url`` is the ``.m3u8`` playlist URL; ``expires_at`` is an ISO-8601\n    UTC timestamp after which the client should refetch.",
        "operationId": "get_track_stream_api_soundcloud_tracks__track_id__stream_get",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Track Id"
            }
          },
          {
            "name": "force_refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/soundcloud/tracks/{track_id}/like": {
      "post": {
        "tags": [
          "soundcloud"
        ],
        "summary": "Like Track",
        "description": "Like a SoundCloud track on behalf of the authenticated user.",
        "operationId": "like_track_api_soundcloud_tracks__track_id__like_post",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Track Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "soundcloud"
        ],
        "summary": "Unlike Track",
        "description": "Unlike a SoundCloud track on behalf of the authenticated user.",
        "operationId": "unlike_track_api_soundcloud_tracks__track_id__like_delete",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Track Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/soundcloud/playlists/{playlist_id}": {
      "delete": {
        "tags": [
          "soundcloud"
        ],
        "summary": "Delete Playlist",
        "description": "Delete a SoundCloud playlist on behalf of the authenticated user.\n\nThe browser can't call ``DELETE /playlists/...`` directly \u2014 SoundCloud\ndoesn't expose that method in its CORS policy \u2014 so this endpoint forwards\nthe request with the user's own token.",
        "operationId": "delete_playlist_api_soundcloud_playlists__playlist_id__delete",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Playlist Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/soundcloud/system-playlists/{urn}/tracks": {
      "get": {
        "tags": [
          "soundcloud"
        ],
        "summary": "Get System Playlist Tracks",
        "description": "Hydrate a system playlist's tracks to full Track payloads.\n\napi-v2 returns slim tracks on the system-playlist resource itself;\nwe ``/tracks?ids=...`` to get full metadata (title, artwork, user).",
        "operationId": "get_system_playlist_tracks_api_soundcloud_system_playlists__urn__tracks_get",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "System playlist URN, e.g. soundcloud:system-playlists:weekly:123",
              "title": "Urn"
            },
            "description": "System playlist URN, e.g. soundcloud:system-playlists:weekly:123"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemPlaylistTracksResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/bpm/soundcloud": {
      "post": {
        "tags": [
          "bpm"
        ],
        "summary": "Save Soundcloud Bpm",
        "description": "Persist a BPM value for a SoundCloud track.",
        "operationId": "save_soundcloud_bpm_api_bpm_soundcloud_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SoundcloudBpmPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoundcloudBpmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/bpm/soundcloud/{track_id}": {
      "get": {
        "tags": [
          "bpm"
        ],
        "summary": "Get Soundcloud Bpm",
        "description": "Return the cached BPM for a SoundCloud track, or 404 if not analyzed.",
        "operationId": "get_soundcloud_bpm_api_bpm_soundcloud__track_id__get",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Track Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SoundcloudBpmResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Soundcloud Bpm Api Bpm Soundcloud  Track Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/suggestions/track": {
      "post": {
        "tags": [
          "suggestions"
        ],
        "summary": "Suggest Track Metadata",
        "description": "Compute ranked metadata suggestions for a single local track.",
        "operationId": "suggest_track_metadata_api_suggestions_track_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/rekordbox/playlists/{playlist_id}/tracks": {
      "get": {
        "tags": [
          "rekordbox"
        ],
        "summary": "Get Playlist Tracks",
        "operationId": "get_playlist_tracks_api_rekordbox_playlists__playlist_id__tracks_get",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Playlist Id"
            }
          },
          {
            "name": "device",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "USB device id; omit for the local install",
              "title": "Device"
            },
            "description": "USB device id; omit for the local install"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TracksResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SoundcloudBpmPayload": {
        "properties": {
          "track_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Track Id"
          },
          "bpm": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Bpm"
          }
        },
        "type": "object",
        "required": [
          "track_id",
          "bpm"
        ],
        "title": "SoundcloudBpmPayload",
        "description": "Client-computed BPM for a SoundCloud track."
      },
      "SoundcloudBpmResponse": {
        "properties": {
          "track_id": {
            "type": "integer",
            "title": "Track Id"
          },
          "bpm": {
            "type": "integer",
            "title": "Bpm"
          }
        },
        "type": "object",
        "required": [
          "track_id",
          "bpm"
        ],
        "title": "SoundcloudBpmResponse"
      },
      "StreamUrlResponse": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "url",
          "expires_at"
        ],
        "title": "StreamUrlResponse",
        "description": "Signed HLS stream URL for a SoundCloud track."
      },
      "SuggestionRequest": {
        "properties": {
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "sc_track": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SCTrackPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "current": {
            "$ref": "#/components/schemas/TrackInfoUpdateRequest"
          }
        },
        "type": "object",
        "required": [
          "file_path"
        ],
        "title": "SuggestionRequest",
        "description": "Request body for the suggestion endpoint.\n\nParameters\n----------\nfile_path:\n    Absolute or root-relative path of the local track being edited. The\n    server uses it for filename parsing and (in the future) reading\n    existing on-disk tags as a suggestion source.\nsc_track:\n    Optional SoundCloud track payload. When ``None``, only filename- and\n    tag-based suggestions are produced.\ncurrent:\n    In-flight editor state. Suggestions equal to a current value are\n    suppressed so the UI never proposes a no-op."
      },
...
      "SystemPlaylistTracksResponse": {
        "properties": {
          "tracks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Tracks"
          }
        },
        "type": "object",
        "required": [
          "tracks"
        ],
        "title": "SystemPlaylistTracksResponse"
      },
...
    }
  }
}